==> morph_tagging/01_finetune_bert_morph_tagger.ipynb <==
# Bert-based morphological tagger <!-- Morfoloogiline ühestaja -->

## Table of contents
[Code](#kood)

   1. [**enc2017 corpus**](#enc_korpus)
      1. [**Defined functions**](#defineeritud_funktsioonid)
      2. [**Collecting data and applying morphological analysis**](#andmete_lugemine_kogumine_ja_margendamine)
         * [Verification](#kontrollimine_1)
      3. [**Statistics**](#statistika)
      4. [**Data splitting**](#andmete_jagamine_treening_test_ja_valideerimishulkadesse)
         * [Grouping data by text type](#andmete_grupeerimine_tekstiliigiti)
         * [Splitting data to sets](#andmete_jagamine_hulkadesse)
      5. [**Token classification model (for assigning morpological categories)**](#ner_mudel)
      6. [*Token classification model (for assigning morpological categories) by text type*](#ner_mudel_tekstiliigiti)

[End](#end)

<!-- ## TODO
1) - [ ] Neurovõrkudel põhineva morfoloogilise ühestaja loomine ja integreerimine EstNLTK teegiga:
    1) - [ ] mudeli treenimiseks vajaliku töövoo loomine;
    2) - [ ] töövoo kasutamine mudeli treenimiseks;
    3) - [ ] mudeli integreerimine EstNLTK teeki.

2) - [ ] Morfoloogilise ühestaja treenimiseks vajalike andmestike moodustamine olemasolevatest andmestikest:
    1) - [ ] morfoloogilisemärgenduse loomine kasutades olemaolevaid EstNLTK automaatmärgendamise töövoogusid;
    2) - [ ] morfoloogilisemärgenduse loomine kasutades UD puudepanga andmeid;

3) - [ ] Saadud ühestaja tulemuste võrdlemine teiste kasutatavate rakendustega:
    1) - [ ] täpsuse ja saagise absoluutne hindamine testvalimil;
    2) - [ ] täpsuse ja saagise relatiivne hindamine märgendamata testandmetel;
    3) - [ ] uute informatiivsete treeningandmete välja valimine.

4) - [ ] Juhendmaterjalide  loomine uue ühestaja kasutamiseks
5) - [ ] Zip the model somehow, https://github.com/estnltk/estnltk/blob/main/estnltk_neural/estnltk_neural/taggers/ner/estbertner_tagger.py -->

<a id='kood'></a>

## Code <!-- Kood -->

import os
import csv
import typing
import itertools
import json
import logging
import warnings
import pkg_resources
import types
import pandas as pd
import sklearn as sk
import numpy as np
import estnltk, estnltk.converters, estnltk.taggers
import torch
import math
from tqdm import tqdm
from simpletransformers.ner import NERModel, NERArgs
from bert_morph_tagger import BertMorphTagger

# Get locally imported modules from current notebook - https://stackoverflow.com/questions/40428931/package-for-listing-version-of-packages-used-in-a-jupyter-notebook - Alex P. Miller
def get_imports():
    
    for name, val in globals().items():
        if isinstance(val, types.ModuleType):
            # Split ensures you get root package, 
            # not just imported function
            name = val.__name__.split(".")[0]

        elif isinstance(val, type):
            name = val.__module__.split(".")[0]
            
        # Some packages are weird and have different
        # imported names vs. system/pip names. Unfortunately,
        # there is no systematic way to get pip names from
        # a package's imported name. You'll have to add
        # exceptions to this list manually!
        poorly_named_packages = {
            "PIL": "Pillow",
            "sklearn": "scikit-learn"
        }
        if name in poorly_named_packages.keys():
            name = poorly_named_packages[name]
            
        yield name
imports = list(set(get_imports()))

# The only way I found to get the version of the root package
# from only the name of the package is to cross-check the names 
# of installed packages vs. imported packages
requirements = []
for m in pkg_resources.working_set:
    if m.project_name in imports and m.project_name!="pip":
        requirements.append((m.project_name, m.version))

for r in requirements:
    print("{}=={}".format(*r))

<a id='enc_korpus'></a>

### enc2017 corpus

Data has been gathered from the [corpus](https://github.com/estnltk/eval_experiments_lrec_2020/blob/master/scripts_and_data/enc2017_selection_plain_texts_json.zip) that was used in evaluation experiments reported in LREC 2020 paper "EstNLTK 1.6: Remastered Estonian NLP Pipeline"[^1].

[^1]: Sven Laur, Siim Orasmaa, Dage Särg, Paul Tammo. "EstNLTK 1.6: Remastered Estonian NLP Pipeline". *Proceedings of The 12th Language Resources and Evaluation Conference*. European Language Resources Association: Marseille, France, May 2020, p. 7154-7162

<a id='defineeritud_funktsioonid'></a>

#### Defined functions <!-- Defineeritud funktsioonid -->

Reading text from files, morphologically tagging texts and save tagged texts file by file into CSV.

def create_csv_file_by_file_enc2017(
    jsons: typing.List[str], 
    in_dir: str, 
    csv_dir: str
    ):
    """
    Creates a CSV file for each text file. \n
    Skips CSV files that have already been created. \n
    For each <code>.json</code> file, the following info is gathered:
    <ul>
        <li><code>sentence_id</code> -- given for each sentence</li>
        <li><code>words</code> -- words gathered from text</li>
        <li><code>form</code> -- word form notation</li>
        <li><code>pos</code> -- part of speech</li>
        <li><code>type</code> -- text type (i.e. genre)</li>
        <li><code>source</code> -- file name where the text is taken from</li>
    </ul>
    <a href="https://github.com/Filosoft/vabamorf/blob/e6d42371006710175f7ec328c98f90b122930555/doc/tagset.md">Tables of morphological categories</a> for more information about <code>form</code> and <code>pos</code>.

    Args:
        jsons (list): List of json files from which to read in the text
        in_dir (str): Directory where to read the json files from
        csv_dir (str): Directory where to save the new csv files
    """
    print("Beginning to morphologically tag file by file")
    for file_name in tqdm(jsons):
        tokens = list()
        sentence_id = 0

        # Skipping previous CSV files
        csv_file_name = file_name[:-4]+'csv'
        if os.path.exists(os.path.join(csv_dir, csv_file_name)):
            # print(f"Skipping {file_name} as {csv_file_name} already exists.")
            continue

        # print(f"Beginning to tag {file_name}")

        # Morph. tagging using estnltk
        text = estnltk.converters.json_to_text(file=os.path.join(in_dir, file_name))
        text_type = text.meta.get('texttype') # Text type
        if file_name.startswith('wiki17'):
            text_type = 'wikipedia'
        elif file_name.startswith('web13'):
            text_type = 'blogs_and_forums'
        morph_analysis = text.tag_layer('morph_analysis')
        for sentence in morph_analysis.sentences:
            sentence_analysis = sentence.morph_analysis
            for text, form, pos in zip(sentence_analysis.text, sentence_analysis.form, sentence_analysis.partofspeech):
                if text:
                    tokens.append((sentence_id, text, form[0], pos[0], text_type, file_name)) # In case of multiplicity, select the first or index 0
            sentence_id += 1

        # print(f"{file_name} tagged, now saving")

        # Salvestamine
        with open(os.path.join(csv_dir, csv_file_name), 'w') as f:
            fieldnames = ['sentence_id', 'word', 'form', 'pos']
            writer = csv.writer(f, delimiter=',', quotechar='"', quoting=csv.QUOTE_MINIMAL)
            writer.writerow(fieldnames)
            for row in tokens:
                writer.writerow(row)

        # print(f"{file_name} saved to {csv_file_name}\n")

    print("Morphological tagging completed successfully")

Reading text from files, morphologically tagging texts and save tagged texts into one whole dataset. <!-- Failidest teksti lugemine, märgendamine ja salvestamine kokku üheks andmestikuks -->

def create_df_enc2017_csv(
    jsons: typing.List[str], 
    in_dir: str, 
    csv_filename: str
    ):
    """
    Creates a new dataset from converted the Estonian UD EDT <a href="https://github.com/UniversalDependencies/UD_Estonian-EDT">corpus</a>. \n
    For each <code>.json</code> file, the following info is gathered:
    <ul>
        <li><code>sentence_id</code> -- given for each sentence</li>
        <li><code>words</code> -- words gathered from text</li>
        <li><code>form</code> -- word form notation</li>
        <li><code>pos</code> -- part of speech</li>
        <li><code>file_prefix</code> -- metadata</li>
        <li><code>source</code> -- file name where the text is taken from</li>
    </ul>
    <a href="https://github.com/Filosoft/vabamorf/blob/e6d42371006710175f7ec328c98f90b122930555/doc/tagset.md">Tables of morphological categories</a> for more information about <code>form</code> and <code>pos</code>.

    Args:
        jsons (list): List of json files from which to read in the text
        in_dir (str): Directory containing list of files (<code>jsons</code>)
        csv_filename (str): CSV filename where to save the gathered text
    """
    tokens = list()
    sentence_id = 0
    fieldnames = ['sentence_id', 'words', 'form', 'pos', 'type', 'source']

    print("Beginning to morphologically tag file by file. This can take a while.")
    for file_name in tqdm(jsons):
        # print(f"Beginning to tag {file_name}")
        sentence_id = 0

        # Morph. tagging
        text = estnltk.converters.json_to_text(file=os.path.join(in_dir, file_name))
        text_type = text.meta.get('texttype') # Text type
        if file_name.startswith('wiki17'):
            text_type = 'wikipedia'
        elif file_name.startswith('web13'):
            text_type = 'blogs_and_forums'
        morph_analysis = text.tag_layer('morph_analysis')
        for sentence in morph_analysis.sentences:
            sentence_analysis = sentence.morph_analysis
            for text, form, pos in zip(sentence_analysis.text, sentence_analysis.form, sentence_analysis.partofspeech):
                if text:
                    tokens.append((sentence_id, text, form[0], pos[0], text_type, file_name)) # In case of multiplicity, select the first or index 0
            sentence_id += 1
        # print(f"{file_name} tagged")

    print("Morphological tagging completed successfully")
    print("Creating Pandas dataframe")
    df = pd.DataFrame(data=tokens, columns=fieldnames)
    df.to_csv(path_or_buf=csv_filename, index=False)
    print(f"Tagged texts saved to {csv_filename}\n")

Cleaning dataset by:
    <ul>
        <li>filling NaN values in columns <code>form</code> and <code>pos</code> with empty strings;</li>
        <li>removing NaN words.</li>
    </ul>

def clean_df(
    df: pd.DataFrame, 
    df_file_name: typing.Optional[str] = None
    ):
    """Finishes dataframe by:
    <ul>
        <li>filling NaN values in columns <code>form</code> and <code>pos</code> with empty strings;</li>
        <li>removing NaN words.</li>
    </ul>

    Args:
        df (pandas.DataFrame): Pandas dataframe to clean
        df_file_name (str): CSV file name from which dataframe was created
    """
    print("Assigning NaN values in columns form and pos with an empty string")
    # NaN values are assigned with an empty string
    df['form'] = df['form'].fillna('')
    df['pos'] = df['pos'].fillna('')
    print("Removing NaN words")
    # Removing NaN words
    df.dropna(subset=['words'], inplace=True)
    if df_file_name:
        df.to_csv(path_or_buf=df_file_name, index=False)
        print(f"Modified dataframe saved to {df_file_name}")
    else:
        print("Dataframe cleaned")

Creation of a new column `labels` concatenating the values of columns `pos` (part of speech) and `form` (word form notation)<!-- Uue veeru `labels` loomine, mis konkateneerib veergude `pos`(sõnaliik) ja `form`(sõnavormi tähistus) väärtused -->

# New 'labels' column
def create_labels_column(
    df: pd.DataFrame, 
    df_file_name: typing.Optional[str] = None
    ):
    """
    Creates a new column <code>labels</code> concatenating the values of columns <code>pos</code> (part of speech) and <code>form</code> (word form notation)

    Args:
        df (pd.DataFrame): Pandas dataframe to create a new column
        df_file_name (str): CSV file name from which dataframe was created
    """
    print("Creating column 'labels'")
    df['labels'] = df.apply(lambda row: str(row['form']) + '_' + str(row['pos']) if row['form'] and row['pos'] else str(row['form']) or str(row['pos']), axis=1)
    print("Column 'labels' created")
    if df_file_name:
        df.to_csv(path_or_buf=df_file_name, index=False)
        print(f"Modified dataframe saved to {df_file_name}")

Obtaining unique tags that the model must predict<!-- Unikaalsete väärtuste saamine, mida mudel peab hakkama ennustama -->

def get_unique_labels():
    """Creates list of unique labels that the model must predict by creating all possible combinations of POS (Part Of Speech) and form.

    <i>Gathering unique labels from the enc2017 database proved to be insufficient for future model evaluation,
    because the database does not contain all possible combinations of POS and form.
    Evaluating model with UD Est-EDT test corpus proved that this problem existed.</i>

    Returns:
        list: List of unique labels
    """
    # Separately, if one of two doesn't exist
    pos_labels = ['A', 'C', 'D', 'G', 'H', 'I', 'J', 'K', 'N', 'O', 'P', 'S', 'U', 'V', 'X', 'Y', 'Z']
    form_labels = ['ab', 'abl', 'ad', 'adt', 'all', 'el', 'es', 'g', 'ill', 'in', 'kom', 'n', 'p', 'pl', 'sg', 'ter', 'tr', 'b', 'd', 'da', 'des', 'ge', 'gem', 'gu', 'ks', 'ksid', 'ksime', 'ksin', 'ksite', 'ma', 'maks', 'mas', 'mast', 'mata', 'me', 'n', 'neg', 'nud', 'nuks', 'nuksid', 'nuksime', 'nuksin', 'nuksite', 'nuvat', 'o', 's', 'sid', 'sime', 'sin', 'site', 'ta', 'tagu', 'taks', 'takse', 'tama', 'tav', 'tavat', 'te', 'ti', 'tud', 'tuks', 'tuvat', 'v', 'vad', 'vat']

    pos_labels_mutable = ['A', 'C', 'N', 'H', 'O', 'P', 'S', 'U', 'Y', 'X']
    #pos_labels_immutable = ['D', 'G', 'I', 'J', 'K', 'X', 'Y', 'Z']
    pos_label_verb = ['V']
    form_labels_noun = ['ab', 'abl', 'ad', 'adt', 'all', 'el', 'es', 'g', 'ill', 'in', 'kom', 'n', 'p', 'ter', 'tr']
    form_labels_noun_count = ['pl', 'sg']
    form_labels_verb = ['b', 'd', 'da', 'des', 'ge', 'gem', 'gu', 'ks', 'ksid', 'ksime', 'ksin', 'ksite', 'ma', 'maks', 'mas', 'mast', 'mata', 'me', 'n', 'neg', 'neg ge', 'neg gem', 'neg gu', 'neg ks', 'neg me', 'neg nud', 'neg nuks', 'neg o', 'neg vat', 'neg tud', 'nud', 'nuks', 'nuksid', 'nuksime', 'nuksin', 'nuksite', 'nuvat', 'o', 's', 'sid', 'sime', 'sin', 'site', 'ta', 'tagu', 'taks', 'takse', 'tama', 'tav', 'tavat', 'te', 'ti', 'tud', 'tuks', 'tuvat', 'v', 'vad', 'vat']

    noun_labels_without_pos = list(itertools.product(form_labels_noun_count, form_labels_noun))
    noun_labels_nested = list(itertools.product(noun_labels_without_pos, pos_labels_mutable))
    form_pos_labels = list(itertools.product(form_labels, pos_labels))
    noun_labels = list()
    verb_labels = list(itertools.product(form_labels_verb, pos_label_verb))

    # Connect count and form in mutables
    for form, pos in noun_labels_nested:
        noun_labels.append((form[0] + ' ' + form[1], pos))

    pos_label_only = [('', pos) for pos in pos_labels]
    form_label_only = [(form, '') for form in form_labels]
    unknown_form_labels = [('?', pos) for pos in pos_labels] # form '?' comes from enc2017 corpus after tagging

    unique_labels = pos_label_only + form_label_only + noun_labels + verb_labels + unknown_form_labels + form_pos_labels + ['?'] # '?' for labels unknown to Vabamorf

    unique_labels_df = pd.DataFrame(unique_labels, columns=['form', 'pos'])
    create_labels_column(unique_labels_df)
    unique_labels_df.drop(labels=['form', 'pos'], axis=1)
    print("List of unique labels created")
    return unique_labels_df['labels'].tolist()

A function that collects texts for each text type *more or less* in proportion to the number of words given as `n` <!-- Funktsioon, millega kogutakse iga tekstiliigi kohta tekste enam-vähem proportsionaalselt sõnade arvu suhtes -->

# Function to collect texts file by file
def gather_rows_for_text_type(
    df: pd.DataFrame, 
    n: int, 
    random_state: typing.Optional[int] = None
    ):
    """Gathers about `n` (>= n) rows for each text type\n
    Ensures that all text types have about the same number of words.

    Args:
        df (pd.DataFrame): The DataFrame containing the text data.
        n (int): Number of words to gather.
        random_state (optional, int): Seed for the shuffle function (acts like <code>random_state</code>).

    Returns:
        pd.DataFrame: Gathered rows for each text type.
    """

    def gather_rows_for_type(
        group: pd.Group, 
        n: int, 
        random_state: typing.Optional[int] = None
        ):
        """Gathers about `n` (>= n) rows for the text type\n

        Args:
            group (): Pandas dataframe group
            n (int): Number of words to gather
            random_state (int): Seed for the shuffle function (acts like <code>random_state</code>)

        Returns:
            pd.DataFrame: Gathered rows for text type
        """
        gathered_rows = pd.DataFrame()
        sources = group['source'].unique()

        if random_state:
            np.random.seed(random_state)
            np.random.shuffle(sources)

        for source in sources:
            source_rows = group[group['source'] == source]
            gathered_rows = pd.concat([gathered_rows, source_rows])
            if len(gathered_rows) >= n:
                break

        return gathered_rows

    grouped = df.groupby('type')
    data = pd.concat([gather_rows_for_type(group, n, random_state) for _, group in grouped])
    return data

Function to predict tags to words

def predict_tags(model, sentences):
    """Predicts tags to a sentence or number of sentences

    Args:
        model: Model to use to predict tags.
        sentences (List[str] or str): List of sentences strings or a sentence string

    Raises:
        TypeError: Variable <code>sentences</code> is not List[str] or str

    Returns:
        (predictions, raw_outputs): Predictions and model's raw output
    """

    if isinstance(sentences, str):
        text = estnltk.Text(sentences)
        text.tag_layer("morph_analysis")
        text.morph_analysis;
        sentences = [s.text for s in text.sentences]
        predictions, raw_outputs = model.predict(sentences, split_on_space=False)

    elif isinstance(sentences, list) and isinstance(sentences[0], list):
        predictions, raw_outputs = model.predict(sentences, split_on_space=False)

    else:
        raise TypeError(f"Input is in wrong format. Possible formats are str or list of lists. Your input is {type(sentences)}")

    return predictions, raw_outputs

Function for predicting top n tags to words

def get_top_n_tags(raw_outputs, tag_list, n=3, with_confidence=True):
    """Extract the top <code>n</code> tags and their probabilities for each word based on the raw output logits.

    Args:
        raw_outputs (list): Raw prediction logits from the model.
        tag_list (list): List of all possible tags.
        n (int): Number of top tags to extract.

    Returns:
        list: A list of lists, where each sublist contains tuples of the top `n` tags and their probabilities for a word.
    """
    top_n_tags = []

    for sentence_logits in raw_outputs:
        for word_logits in sentence_logits:

            # Get the logits for the word
            logits = list(word_logits.values())[0][0]

            # Apply softmax to get probabilities
            probabilities = torch.nn.functional.softmax(torch.tensor(logits), dim=-1).numpy()

            # Get the indices of the top `n` probabilities
            top_n_indices = np.argsort(probabilities)[-n:][::-1]

            # Map these indices to the actual tags and their probabilities
            if with_confidence:
                top_n_tags.append(
                    [(tag_list[i], probabilities[i]) for i in top_n_indices]
                )
            else:
                top_n_tags.append(
                    [(tag_list[i]) for i in top_n_indices]
                )

    return top_n_tags

def predict_top_n_tags(model, to_predict, with_confidence=True):
    """Predicts top <code>n</code> tags and their probabilities for each word.

    Args:
        to_predict (list): A python list of text (str) to be sent to the model for prediction.
        Must be a a list of lists, with the inner list being a list of strings consisting of the split sequences. 
        The outer list is the list of sequences to predict on.
        with_confidence (bool): Whether to output confidence values for each tag.
    """
    _, raw_outputs = model.predict(to_predict, split_on_space=False)

    tag_list = model.config.id2label  # Assuming the model has a mapping from ids to labels
    top_n_tags_per_word = get_top_n_tags(raw_outputs, tag_list, n=3, with_confidence=with_confidence)

    for sentence in to_predict:
        for word, tags in zip(sentence, top_n_tags_per_word):
            print(f"Word: {word}")
            if with_confidence:
                for tag, confidence in tags:
                    print(f"\tTag: {tag} \t Confidence: {confidence:.4f}")
            else:
                for tag in tags:
                    print(f"\tTag: {tag}")

Functions to predict labels to the group of sentences. Includes sentence splitting into clauses and clauses into equal length clauses' parts.

def check_token_count(model, sentence):
    """Checks token count in the sentence

    Args:
        model (): NER model that is predicted with
        sentence (list): list of words in a sentence

    Returns:
        bool: Whether the token count exceeds model's maximum sequence length
    """

    inputs = model.tokenizer('  '.join(sentence), return_tensors="pt")
    return bool(len(inputs["input_ids"][0]) >= model.args.max_seq_length)

def get_clause_parts(model, clause):
    """Splits clause into equal length clause segments. <i>Clauses can get long when there is a list in a sentence</i>

    Args:
        model (): NER model that is predicted with
        clause (list): list of words in a clause

    Returns:
        list: List of clause segments. Each segment is a list of words.
    """

    inputs = model.tokenizer('  '.join(clause), return_tensors="pt")
    clause_parts = np.array_split(clause, math.ceil(len(inputs["input_ids"][0]) / model.args.max_seq_length))
    return clause_parts

def get_clauses(model, sentence):
    """Splits sentence into clauses using EstNLTK clauses layer.

    Args:
        model (): NER model that is predicted with
        sentence (list): list of words in a sentence

    Returns:
        list: List of clauses. Each clause is a list of words.
    """

    sentence_text = '  '.join(sentence)
    text = estnltk.Text(sentence_text)
    text.tag_layer('clauses')
    clauses = list()
    for clause in text.clauses:
        if check_token_count(model, clause.text):
            clause_parts = get_clause_parts(model, clause.text)
            clauses.extend(clause_parts)
        else:
            clauses.append(clause.text)
    return clauses

def get_clauses_labels(model, sentence):
    """Predicts labels to clauses

    Args:
        model (): NER model that is predicted with
        sentence (list): list of words in a sentence

    Returns:
        list: List of tags predicted to words in clauses
    """

    clauses = get_clauses(model, sentence)

    ner_labels_parts = list()

    # Predict tags
    predictions, raw_outputs = model.predict(clauses, split_on_space=False)
    for prediction_part in predictions:
        # ner_words_part = [list(p.keys())[0] for p in prediction_part]
        ner_labels_part = [list(p.values())[0] for p in prediction_part]

        # ner_words_parts.append(ner_words_part)
        ner_labels_parts.append(ner_labels_part)

    # ner_words = list(itertools.chain.from_iterable(ner_words_parts))
    ner_labels = list(itertools.chain.from_iterable(ner_labels_parts))

    return ner_labels

def get_sentence_labels(model, sentence):
    """Predicts labels to a sentence

    Args:
        model (): NER model that is predicted with
        sentence (list): list of words in a sentence

    Returns:
        list: List of tags predicted to words in a sentence
    """

    # Predict tags
    predictions, raw_outputs = model.predict([sentence], split_on_space=False)

    # ner_words = [list(p.keys())[0] for p in predictions[0]]
    ner_labels = [list(p.values())[0] for p in predictions[0]]
    return ner_labels

def process_groups(model, groups):
    """Predicts labels to a list of groups. This group contains sentences for each source text file.

    Args:
        model (): NER model that is predicted with
        groups (): Group containing sentences for each source text file

    Raises:
        AssertionError: When the length of predicted labels mismatch the length of the sentence length. 
        <i>This might mean that generated clauses are wrong (meaning some words are missing because of <code>'  '.join(sentence)</code>)</i>

    Returns:
        list: List of predicted labels for each sentence in each source text file
    """
    chunk_results = []

    for _, group in groups:
        sentence = group.words.tolist()

        if check_token_count(model, sentence):  # Sentence splitting if token count is above model's max sequence length
            ner_labels = get_clauses_labels(model, sentence)
        else:
            ner_labels = get_sentence_labels(model, sentence)

        if len(ner_labels) != len(sentence):
            display(group)
            print(len(sentence), len(ner_labels))
            raise AssertionError("Predicted labels length mismatch sentence length.")

        group['ner_labels'] = ner_labels
        chunk_results.append(group)

    return chunk_results

<a id='andmete_lugemine_kogumine_ja_margendamine'></a>

#### Collecting data and applying morphological analysis <!-- Andmete lugemine kogumine ja margendamine -->

Reading text from files, morphologically tagging texts and save them file by file. <!--Failidest teksti lugemine, märgendamine ja salvestamine failhaaval -->

Collecting `.json` files from directory <!--`json` failide kogumine kaustast-->

in_dir = '_plain_texts_json'

jsons = os.listdir(in_dir)

Data creation and cleaning

if not os.path.exists('./andmestik.csv'):
    create_df_enc2017_csv(jsons, in_dir, 'andmestik.csv')

csv_file = "andmestik.csv"
df = pd.read_csv(csv_file, keep_default_na=False)
clean_df(df, csv_file)

create_labels_column(df, csv_file)

<a id='kontrollimine_1'></a>

##### Verification <!-- Kontrollimine -->

Ensuring that correct info was collected from the json file<!-- Kontrollitakse, et json failist said õiged infokillud kogutud -->

# For JSON file
text = estnltk.converters.json_to_text(file='.\\_plain_texts_json\\nc_255_27981.json')
morph_analysis = text.tag_layer('morph_analysis')
ma_json = morph_analysis.sentences[0].morph_analysis
test_json = pd.DataFrame(data=list(zip(ma_json.text, ma_json.form, ma_json.partofspeech)), columns=("text", "form", "partofspeech"))
# display(test_json)

# For CSV file
lines = list()

with open('.\\_plain_texts_csv\\nc_255_27981.csv', 'r') as f:
    csvFile = csv.reader(f, delimiter=',', quotechar='"', quoting=csv.QUOTE_MINIMAL)
    for line in csvFile:
        if line:
            lines.append(line)

test_csv = pd.DataFrame(lines[1:], columns=lines[0])

# display(test_csv)

# For Pandas dataframe
test_csv = df[df['source'] == 'nc_255_27981.json']

Example of the first sentence in file `nc_255_27981`<!-- Näide esimesest lausest failis `nc_255_27981` -->

display(test_json[-15:])
display(test_csv[test_csv.sentence_id == 0][-15:])

<a id='statistika'></a>

#### Statistics <!-- Statistika -->

# Reading CSV file
csv_file = "andmestik.csv"
df = pd.read_csv(csv_file, keep_default_na=False)

Token count in each text type<!-- Sõnade arv igal tekstiliigil -->

print("Word count")
display(df['type'].value_counts())
print(f'Sum: {df['type'].count()}')

Sentence count in each text type<!-- Lausete arv igal tekstiliigil -->

text_types = df['type'].unique().tolist()
print("Sentence count")
for text_type in text_types:
    sentence_grouped = df[df['type'] == text_type].groupby(['source', 'sentence_id'])
    sentence_count = sentence_grouped.ngroups
    print(f"{text_type}: {sentence_count}")
sentence_grouped = df.groupby(['source', 'sentence_id'])
sentence_count = sentence_grouped.ngroups
print(f"\nSum: {sentence_count}")

File count in each text type<!-- Failide arv tekstiliigiti -->

print("File count")
for text_type in text_types:
    print(f"{text_type}: {len(df['source'][df['type'] == text_type].value_counts().tolist())}")
print(f"\nSum: {len(df['source'].value_counts().tolist())}")

Outputs word count for each source (i.e. text file) present in the data

def check_each_source_word_count(data):
    """Outputs word count for each source (i.e. text file) present in the data

    Args:
        data (pd.DataFrame): Pandas dataframe
    """
    for text_type in text_types:
        print(f"{text_type}, sources: {len(data['source'][data['type'] == text_type].value_counts().tolist())}")
        display(data['source'][data['type'] == text_type].value_counts().head(5))
        if len(data['source'][data['type'] == text_type].value_counts().tolist()) > 5:
            display(data['source'][data['type'] == text_type].value_counts().tail(5))

# check_each_source_word_count(df)

<a id='andmete_jagamine_treening_test_ja_valideerimishulkadesse'></a>

#### Data splitting<!-- ### Andmete jagamine treening-, test- ja valideerimishulkadesse -->

The size of the data used in training was based on the [statistics](https://github.com/UniversalDependencies/UD_Estonian-EDT/blob/master/stats.xml) found in [Estonian UD EDT treebank](https://github.com/UniversalDependencies/UD_Estonian-EDT). Data is gathered in such a way that all text types have more or less proportionate numbers of words.

|Text type|Word count|
|---|---|
|Fiction|             115847|
|Blogs and forums|    115719|
|Science|             115347|
|Periodicals|         115050|
|Wikipedia|           115014|
|**Sum**|             576977|

The tag that the model starts to predict consists of a concatenation of columns `form` and `pos` connected with underscore _ symbol (e.g. `form` s and `pos` V will create s_V). In the data those concatenations will be stored in the column `labels`.

<a id='andmete_grupeerimine_tekstiliigiti'></a>

##### Grouping data by text type <!-- #### Andmete grupeerimine tekstiliigiti -->

# Reading CSV file
csv_file = "andmestik.csv"
df = pd.read_csv(csv_file, keep_default_na=False)

```
#pos_labels = ['A', 'C', 'D', 'G', 'H', 'I', 'J', 'K', 'N', 'O', 'P', 'S', 'U', 'V', 'X', 'Y', 'Z']
#form_labels = ['ab', 'abl', 'ad', 'adt', 'all', 'el', 'es', 'g', 'ill', 'in', 'kom', 'n', 'p', 'pl', 'sg', 'ter', 'tr', 'b', 'd', 'da', 'des', 'ge', 'gem', 'gu', 'ks', 'ksid', 'ksime', 'ksin', 'ksite', 'ma', 'maks', 'mas', 'mast', 'mata', 'me', 'n', 'neg', 'neg ge', 'neg gem', 'neg gu', 'neg ks', 'neg me', 'neg nud', 'neg nuks', 'neg o', 'neg vat', 'neg tud', 'nud', 'nuks', 'nuksid', 'nuksime', 'nuksin', 'nuksite', 'nuvat', 'o', 's', 'sid', 'sime', 'sin', 'site', 'ta', 'tagu', 'taks', 'takse', 'tama', 'tav', 'tavat', 'te', 'ti', 'tud', 'tuks', 'tuvat', 'v', 'vad', 'vat']
```

Labels that have '?' in them

question_mark_labels = df[df['labels'].str.contains(r"\?")]
question_mark_labels['labels'].unique()

unique_labels = get_unique_labels()
# display(unique_labels)

Saving unique labels for future use

if not os.path.exists("./unique_labels.json"):
    with open("unique_labels.json", 'w') as f:
        json.dump(unique_labels, f, indent=2)

data = gather_rows_for_text_type(df, 115000) #115000 words per text type for NER model
display(data)

# Result
display(data['type'].value_counts())
print(f"Sum: {data['words'].size}")

Gathers text types for evaluating model by text type

text_types = data['type'].unique().tolist()

Saving model data to CSV file

data.to_csv(path_or_buf='model_data.csv', index=False)

<a id='andmete_jagamine_hulkadesse'></a>

##### Splitting data to sets<!-- #### Andmete jagamine hulkadesse -->

# Reading CSV file
data = pd.read_csv("model_data.csv", keep_default_na=False)

Grouping data by filename to preserve the integrity of texts<!-- Andmete grupeerimine failinimede kaupa, et säilitada tekstide terviklikkus -->

grouped = data.groupby('source')

groups = list(grouped.groups.keys())
train_groups, test_groups = sk.model_selection.train_test_split(groups, test_size=0.2, random_state=42)

def filter_by_group(df, groups):
    return df[df['source'].isin(groups)]

# Splitting dataframe
train_df = filter_by_group(data, train_groups)
test_df = filter_by_group(data, test_groups)

Removing unnecessary columns for the model<!-- Mudelile ebavajalike veergude eemaldamine -->

train_df = train_df.drop(labels=['type', 'source'], axis=1)
test_df = test_df.drop(labels=['type', 'source'], axis=1)
# display(train_df)
# display(test_df)

<a id='ner_mudel'></a>

#### Token classification model (for assigning morpological categories)<!-- ### NER Mudel -->

with open("unique_labels.json", 'r') as f:
    unique_labels = json.load(f)

Initializing the model<!-- Mudeli ülesehitamine -->

# Set up logging
logger = logging.getLogger('simpletransformers.ner.ner_model')
logger.setLevel(logging.ERROR)

# Suppress specific warnings
# warnings.filterwarnings("ignore", category=FutureWarning) # For warning message "FutureWarning: `torch.cuda.amp.autocast(args...)` is deprecated."
warnings.filterwarnings("ignore", category=UserWarning) # For warnings from seqeval.metrics like "UserWarning: <tag> seems not to be NE tag."

# Configurations
model_args = NERArgs()
model_args.train_batch_size = 8
model_args.evaluate_during_training = False
model_args.learning_rate = 5e-5
model_args.num_train_epochs = 10
model_args.use_early_stopping = True
model_args.use_cuda = torch.cuda.is_available()  # Use GPU if available
model_args.save_eval_checkpoints = False
model_args.save_model_every_epoch = False # Takes a lot of storage space
model_args.save_steps = -1
model_args.overwrite_output_dir = True
model_args.cache_dir = 'NER_mudel/cache'
model_args.best_model_dir = 'NER_mudel/best_model'
model_args.output_dir = 'NER_mudel'
model_args.use_multiprocessing = False
model_args.silent = True

# Initialization
model = NERModel("camembert", "EMBEDDIA/est-roberta", args=model_args, labels=unique_labels)

Training the model<!-- Mudeli treenimine -->

if not os.path.exists(".\\NER_mudel\\config.json"):
    # Train model
    print("Training model")
    model.train_model(train_df)
else:
    model = NERModel("camembert", "NER_mudel", args=model_args, labels=unique_labels)

Evaluating the model<!-- Mudeli hindamine -->

# Evaluate the model
result, model_outputs, preds_list = model.eval_model(test_df)
print(result)

print(f"Evaluation Loss:{result['eval_loss']:.4f}")
print(f"Precision: \t{result['precision']:.4f}")
print(f"Recall: \t{result['recall']:.4f}")
print(f"F1 Score: \t{result['f1_score']:.4f}")

<!-- Previously, you could see a lot of UserWarnings logged in the output. It did not impose a problem because the model could still predict these tags. This is now fixed by configuring the logger. -->

Evaluating the model by text type<!-- Mudeli hindamine tekstiliigiti -->

for text_type in text_types:

    print(f"Text type:\t{text_type}")

    text_type_data = data.loc[data['type'] == text_type]

    print(f"Token count:\t{len(text_type_data)}")

    print(f"Starting evaluation")

    # Evaluate the model
    result, model_outputs, preds_list = model.eval_model(text_type_data)
    # print(result)
    print()
    print(f"Evaluation Loss:{result['eval_loss']:.4f}")
    print(f"Precision: \t{result['precision']:.4f}")
    print(f"Recall: \t{result['recall']:.4f}")
    print(f"F1 Score: \t{result['f1_score']:.4f}")
    print()

Make predictions with the model <!-- Mudelil ennustamine -->

predictions, raw_outputs = model.predict([['A. H. Tammsaare',
 'oli',
 'eesti',
 'kirjanik',
 ',',
 'esseist',
 ',',
 'kultuurifilosoof',
 'ja',
 'tõlkija',
 '.']], split_on_space=False)

print(predictions)

test_text = estnltk.Text("A. H. Tammsaare oli eesti kirjanik, esseist, kultuurifilosoof ja tõlkija. Üksnes autorihüvitis oli 12 431 krooni.")
test_text.tag_layer("morph_analysis")
test_text.morph_analysis;
sentences = [s.text for s in test_text.sentences]

# Predict tags
predictions, raw_outputs = model.predict(sentences, split_on_space=False)

# Output of predictions
print(predictions)

predict_top_n_tags(model, [['A. H. Tammsaare',
 'oli',
 'eesti',
 'kirjanik',
 ',',
 'esseist',
 ',',
 'kultuurifilosoof',
 'ja',
 'tõlkija',
 '.']])

<a id='ner_mudel_tekstiliigiti'></a>

#### Token classification model (for assigning morpological categories) NER model by text type<!-- ### NER mudel tekstiliigiti -->

for text_type in text_types:

    print(f"Text type:\t{text_type}")

    text_type_data = data.loc[data['type'] == text_type]

    print(f"Token count: {len(text_type_data)}")

    grouped = data.groupby('source')

    groups = list(grouped.groups.keys())
    train_groups, test_groups = sk.model_selection.train_test_split(groups, test_size=0.2, random_state=42)

    # Split the dataframe
    train_df = df[df['source'].isin(train_groups)]
    test_df = df[df['source'].isin(test_groups)]

    print(f"Train- and testset created.\nBeginning to train")

    # Model

    logging.basicConfig(level=logging.INFO)
    transformers_logger = logging.getLogger("transformers")
    transformers_logger.setLevel(logging.WARNING)

    # Configure the model
    model_args = NERArgs()
    model_args.train_batch_size = 8
    model_args.evaluate_during_training = False
    model_args.learning_rate = 5e-5
    model_args.num_train_epochs = 10
    model_args.use_early_stopping = True
    model_args.use_cuda = torch.cuda.is_available()  # Use GPU if available
    model_args.overwrite_output_dir = True
    model_args.save_eval_checkpoints = False
    model_args.save_model_every_epoch = False
    model_args.save_steps = -1
    model_args.cache_dir = text_type + '_NER_mudel/cache'
    model_args.best_model_dir = text_type + '_NER_mudel/best_model'
    model_args.output_dir = text_type + '_NER_mudel'

    # Initialize the model
    model = NERModel("camembert", "EMBEDDIA/est-roberta", args=model_args, labels=unique_labels)

    # Train the model
    model.train_model(train_df, eval_data=test_df)

    print(f"Starting evaluation")

    # Evaluate the model
    result, model_outputs, preds_list = model.eval_model(text_type_data)
    print(result)
    print(f"Evaluation Loss:{result['eval_loss']:.4f}")
    print(f"Precision: \t{result['precision']:.4f}")
    print(f"Recall: \t{result['recall']:.4f}")
    print(f"F1 Score: \t{result['f1_score']:.4f}")

New trained and evaluated NER models

| Text type         | Token count | eval_loss            | precision          | recall             | f1_score           |
|-------------------|-------------|----------------------|--------------------|--------------------|--------------------|
| blogs_and_forums  | 115719      | 0.23976334929466248  | 0.9538220130015691 | 0.9520293551707164 | 0.9529248409925647 |
| fiction           | 115847      | 0.2396537810564041   | 0.9538792523867151 | 0.952342596321654  | 0.9531103049845493 |
| periodicals       | 115050      | 0.24317898601293564  | 0.9550244383659926 | 0.9530585760952253 | 0.9540404945350295 |
| science           | 115347      | 0.2376849427819252   | 0.9557577659240665 | 0.9541325457555824 | 0.9549444643496955 |
| wikipedia         | 115014      | 0.23429438471794128  | 0.9547409930094999 | 0.953416565982011  | 0.9540783198620781 |

General NER model evaluations by text type

| Text type         | Token count | eval_loss            | precision          | recall             | f1_score           |
|-------------------|-------------|----------------------|--------------------|--------------------|--------------------|
| blogs_and_forums  | 115719      | 0.08935269061475992  | 0.974046101995129  | 0.9733038348082595 | 0.9736748269375007 |
| fiction           | 115847      | 0.004841572717996314 | 0.9990726034260374 | 0.9989767640162013 | 0.9990246814225794 |
| periodicals       | 115050      | 0.09677119304736455  | 0.9807560725282244 | 0.9801692452346354 | 0.9804625710743448 |
| science           | 115347      | 0.1642883848398924   | 0.9677601809954751 | 0.9688169868554095 | 0.9682882955716798 |
| wikipedia         | 115014      | 0.2906533405184746   | 0.9479240975888144 | 0.9471808396892152 | 0.9475523228865832 |

General NER model evaluation on test set

|    Metric     |   Value |
|---------------|---------|
|Evaluation Loss|   0.2393|
|Precision| 	    0.9545|
|Recall| 	        0.9534|
|F1 Score| 	        0.9540|

<a id='end'></a>

### END
